pyproject.toml
[project]
name = "esa-mne-challenge"
version = "0.1.0"
description = "Repository for ESA MNE discovery and extraction challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "crawl4ai>=0.6.2",
    "duckduckgo-search>=8.0.1",
    "googlesearch-python>=1.3.0",
    "openai>=1.76.2",
    "pandas>=2.2.3",
    "s3fs>=2025.3.2",
    "tqdm>=4.67.1",
]
authors = [
  {name="Thomas Faria", email="thomas.faria@insee.fr"}
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "pre-commit>=4.2.0",
]

[tool.setuptools]
packages = ["src"]

[tool.ruff]
line-length = 120

[tool.uv]
default-groups = ["dev"]
